app/commands.c app/crop.c app/file_new_dialog.c app/info_dialog.[ch]
1999-05-31  Michael Natterer  <mitschel@cs.tu-berlin.de>

	* app/commands.c
	* app/crop.c
	* app/file_new_dialog.c
	* app/info_dialog.[ch]
	* app/interface.c
	* app/layers_dialog.c
	* app/resize.[ch]
	* app/rotate_tool.c
	* app/scale_tool.c
	* app/shear_tool.c: finished the float->double migration for
	resolution values. Standardized the order of function calls which
	initialize sizeentries. Fixed some off-by-one errors by using
	correct double->int casting. Use the g* counterparts of int and
	double in some places. Various code cleanups.

	* app/preferences_dialog.c: same changes as above plus a cleaner
	implementation of the mem_size_unit stuff. The whole dialog should
	behave like before.

+++ app/tools/gimpsheartool.c
@@ -42,9 +42,9 @@
 #define MIN_MOVE       5
 
 /*  variables local to this file  */
-static int         direction_unknown;
-static float       xshear_val;
-static float       yshear_val;
+static gint        direction_unknown;
+static gdouble     xshear_val;
+static gdouble     yshear_val;
 
 /*  forward function declarations  */
 static void *      shear_tool_recalc  (Tool *, void *);
